+++ zarrs/src/array/codec/array_to_bytes/pcodec/pcodec_partial_decoder.rs
@@ -10,7 +10,7 @@ use crate::{
         },
         ArraySize, ChunkRepresentation, DataType,
     },
-    array_subset::ArraySubset,
+    indexer::IncompatibleIndexerError,
 };
 
 #[cfg(feature = "async")]
@@ -37,16 +37,14 @@ impl PcodecPartialDecoder {
 
 fn do_partial_decode<'a>(
     decoded: Option<RawBytes<'a>>,
-    indexer: &ArraySubset,
+    indexer: &dyn crate::indexer::Indexer,
     decoded_representation: &ChunkRepresentation,
 ) -> Result<ArrayBytes<'a>, CodecError> {
     let chunk_shape = decoded_representation.shape_u64();
     match decoded {
         None => {
-            let array_size = ArraySize::new(
-                decoded_representation.data_type().size(),
-                indexer.num_elements(),
-            );
+            let array_size =
+                ArraySize::new(decoded_representation.data_type().size(), indexer.len());
             let fill_value =
                 ArrayBytes::new_fill_value(array_size, decoded_representation.fill_value());
             Ok(fill_value)
@@ -60,7 +58,7 @@ fn do_partial_decode<'a>(
                     let decoded_chunk: ArrayBytes = decoded_chunk.into();
                     let bytes_subset = decoded_chunk
                         .extract_array_subset(
-                            &indexer,
+                            indexer,
                             &chunk_shape,
                             decoded_representation.data_type(),
                         )?
@@ -126,9 +124,17 @@ impl ArrayPartialDecoderTraits for PcodecPartialDecoder {
 
     fn partial_decode(
         &self,
-        indexer: &ArraySubset,
+        indexer: &dyn crate::indexer::Indexer,
         options: &CodecOptions,
     ) -> Result<ArrayBytes<'_>, CodecError> {
+        if indexer.dimensionality() != self.decoded_representation.dimensionality() {
+            return Err(IncompatibleIndexerError::new_incompatible_dimensionality(
+                indexer.dimensionality(),
+                self.decoded_representation.dimensionality(),
+            )
+            .into());
+        }
+
         let decoded = self.input_handle.decode(options)?;
         do_partial_decode(decoded, indexer, &self.decoded_representation)
     }
@@ -164,14 +170,15 @@ impl AsyncArrayPartialDecoderTraits for AsyncPCodecPartialDecoder {
 
     async fn partial_decode(
         &self,
-        indexer: &ArraySubset,
+        indexer: &dyn crate::indexer::Indexer,
         options: &CodecOptions,
     ) -> Result<ArrayBytes<'_>, CodecError> {
         if indexer.dimensionality() != self.decoded_representation.dimensionality() {
-            return Err(CodecError::InvalidArraySubsetDimensionalityError(
-                indexer.clone(),
+            return Err(IncompatibleIndexerError::new_incompatible_dimensionality(
+                indexer.dimensionality(),
                 self.decoded_representation.dimensionality(),
-            ));
+            )
+            .into());
         }
 
         let decoded = self.input_handle.decode(options).await?;