+++ zarrs/src/array/codec/array_to_bytes/vlen/vlen_partial_decoder.rs
@@ -2,13 +2,10 @@
 
 use std::{num::NonZeroU64, sync::Arc};
 
-use crate::{
-    array::{
-        array_bytes::extract_decoded_regions_vlen,
-        codec::{ArrayPartialDecoderTraits, BytesPartialDecoderTraits, CodecError, CodecOptions},
-        ArrayBytes, ArraySize, ChunkRepresentation, CodecChain, DataType, FillValue, RawBytes,
-    },
-    array_subset::ArraySubset,
+use crate::array::{
+    array_bytes::extract_decoded_regions_vlen,
+    codec::{ArrayPartialDecoderTraits, BytesPartialDecoderTraits, CodecError, CodecOptions},
+    ArrayBytes, ArraySize, ChunkRepresentation, CodecChain, DataType, FillValue, RawBytes,
 };
 use zarrs_metadata_ext::codec::vlen::{VlenIndexDataType, VlenIndexLocation};
 
@@ -53,7 +50,7 @@ fn decode_vlen_bytes<'a>(
     index_data_type: VlenIndexDataType,
     index_location: VlenIndexLocation,
     bytes: Option<RawBytes>,
-    indexer: &ArraySubset,
+    indexer: &dyn crate::indexer::Indexer,
     fill_value: &FillValue,
     shape: &[u64],
     options: &CodecOptions,
@@ -88,7 +85,7 @@ fn decode_vlen_bytes<'a>(
     } else {
         // Chunk is empty, all decoded regions are empty
         let array_size = ArraySize::Variable {
-            num_elements: indexer.num_elements(),
+            num_elements: indexer.len(),
         };
         Ok(ArrayBytes::new_fill_value(array_size, fill_value))
     }
@@ -105,7 +102,7 @@ impl ArrayPartialDecoderTraits for VlenPartialDecoder {
 
     fn partial_decode(
         &self,
-        indexer: &ArraySubset,
+        indexer: &dyn crate::indexer::Indexer,
         options: &CodecOptions,
     ) -> Result<ArrayBytes<'_>, CodecError> {
         // Get all the input bytes (cached due to CodecTraits::partial_decoder_decodes_all() == true)
@@ -166,7 +163,7 @@ impl AsyncArrayPartialDecoderTraits for AsyncVlenPartialDecoder {
 
     async fn partial_decode(
         &self,
-        indexer: &ArraySubset,
+        indexer: &dyn crate::indexer::Indexer,
         options: &CodecOptions,
     ) -> Result<ArrayBytes<'_>, CodecError> {
         // Get all the input bytes (cached due to CodecTraits::partial_decoder_decodes_all() == true)