+++ zarrs/src/lib.rs
@@ -226,6 +226,7 @@ pub mod array;
 pub mod array_subset;
 pub mod config;
 pub mod group;
+pub mod indexer;
 pub mod node;
 pub mod version;